move some kubernetes specific jobs tests
