+++ config/BUILD.bazel
@@ -4,7 +4,7 @@ filegroup(
     name = "prowjobs",
     srcs = glob([
         "jobs/**/*.yaml",
-    ]),
+    ]) + ["//config/jobs/kubernetes-security:generated-jobs"],
 )
 
 filegroup(
@@ -19,6 +19,7 @@ filegroup(
     srcs = [
         ":package-srcs",
         "//config/jobs/kubernetes-security:all-srcs",
+        "//config/tests/jobs:all-srcs",
     ],
     tags = ["automanaged"],
 )
