+++ config/README.md
@@ -1,9 +1,10 @@
 # Kubernetes Project Configs
 
-This is a central place for generic tools configs for Kubernetes project
+This is a central place for Kubernetes-project specific configs for other tools in this repo.
 
 ### UNDER MIGRATION
 
 Directory structure:
 
 [jobs/](./jobs) : job configs for Kubernetes Prow deployment, potentially testgrid configs as well
+[tests](./tests) : validation tests for the configs
