+++ config/jobs/kubernetes-security/BUILD.bazel
@@ -29,6 +29,12 @@ filegroup(
 )
 
 filegroup(
+    name = "generated-jobs",
+    srcs = glob(["*.yaml"]),
+    visibility = ["//visibility:public"],
+)
+
+filegroup(
     name = "all-srcs",
     srcs = [":package-srcs"],
     tags = ["automanaged"],
