+++ config/tests/jobs/README.md
@@ -1,0 +1,3 @@
+# config/tests
+
+This Directory contains tests for the Kubernetes / upstream test-infra deployment.
